requirements.txt
azure-functions
matplotlib==3.1.1
numpy==1.13.3
pandas==0.25.1
pylab==0.0.2